fix: coverage job
Fix the coverage job so we can add back it on the Gitlab CI.
Made some changes to it, removed test job. So with MRs we only have a
coverage job. I did so because it generates junit and cobertuna report.
Which provides us which better context during MRs we don't need a test
report.